requirements.txt
fal-client>=0.0.23
configparser>=5.0.0
requests>=2.25.0
pillow>=9.0.0 